moving FakeNetworkConnection to common test folder
