+++ Assets/Mirror/Tests/Common/FakeNetworkConnection.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 935b1eb49c500674eaaf88982952a69e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
